Fix coremark-pro benchmark
It turns out the default configuration of CoremarkPro is kind of broken,
due to the number of iterations. For the results to be valid each
benchmark needs to run 10s according to the official documentation, but
by default this is not the case at all (some take 1ms for an iteration,
other 500ms, but they all have the same number of iterations!).
I selected a per-benchmark number of iteration that ensures 10+ seconds
of benchmarks, and re-ran the benchmark for coremark native (and Themis
with 8 cores).